implemented clear, erase and some debug operations
sub_protect_overflow is used to assert A - B = 0 where A < B.

and it is called in my_string's member initializer so my_string's size is not overflowed.

However, the C++ Standard doesn't suppose the constructors to indure overflow but to throw an exception when the overflow occured.

So, the constructors have been modified to throw std::out_of_range when the size is to overflow.

CS-1 CS-113 CS-143