specified use of this pointer
In general, this-> is implied when accessing to class members.

However, a class template may have a specilization whose member declarations are different from the template.

So, some compilers(includes gcc) requires explicit use of this pointer when accessing member of class template.

reference: https://stackoverflow.com/questions/993352/when-should-i-make-explicit-use-of-the-this-pointer

CS-1 CS-113 CS-143