added parameter to sample method
Sampler was intended to meet different scale of matrix.
SampleContext has been the mean to describe how to handle with different scale of matrix. However, it was not delivered by the user but was created for its own use, that is, it was not furfilling its purpose.
To acheive the intention, added SampleContext parameter to top level method.

added return value for postManipulate due to adjust method chaining.

setSrc has been integrated into sample method in order to subtle misuse caused by remaining internal state.

CE-17
CE-15
CE-16
CE-18